requirements.txt
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
scikit-learn>=1.3.0
jupyter>=1.0.0
pytest>=7.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
tqdm>=4.66.0
PyYAML>=6.0.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
evaluate>=0.4.0
rouge-score>=0.1.2
nltk>=3.8.0
